Added UploadArgs as memeber of PostArgs
