--- src/models/args/MediaArgs.ts
@@ -1,5 +1,0 @@
-// PACKAGES
-import { ArrayMaxSize, IsArray, IsNotEmpty, IsNumberString, validateSync } from 'class-validator';
-
-// MODELS
-import { DataValidationError } from '../errors/DataValidationError';
